+++ sakura_core/dlg/CDlgFind.cpp
@@ -65,7 +65,7 @@ BOOL CDlgFind::OnCbnDropDown( HWND hwndCtl, int wID )
 	case IDC_COMBO_TEXT:
 		if ( ::SendMessage(hwndCtl, CB_GETCOUNT, 0L, 0L) == 0) {
 			const auto& keys = m_pShareData->m_sSearchKeywords.m_aSearchKeys;
-			for (size_t i = 0; i < keys.size(); ++i) {
+			for (int i = 0; i < keys.size(); ++i) {
 				ApiWrap::Combo_AddString( hwndCtl, keys[i] );
 			}
 		}
